Deleted colliderComponent, Changed position data type to float from int

+++ include/Components/PositionComponent.h
@@ -5,15 +5,15 @@
 
 class PositionComponent : public Component {
 public:
-    int x, y;
+    float x, y;
 
     PositionComponent() { x = 0.0f; y = 0.0f;}
     PositionComponent(float x, float y) : x(x), y(y) {}
     PositionComponent(const Vector2D& pos) : x(pos.x) , y(pos.y) {}
 
     
-    void set(int xVal, int yVal) { x = xVal; y = yVal; }
+    void set(float xVal, float yVal) { x = xVal; y = yVal; }
     void set(const Vector2D& pos) { x = pos.x; y = pos.y; }
-    void add(int xVal, int yVal) { x += xVal; y += yVal; }
+    void add(float xVal, float yVal) { x += xVal; y += yVal; }
     void add(const Vector2D& pos) { x += pos.x; y += pos.y;}
 };
